\name{revcode}
\alias{revcode}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Reverse coding variables
}
\description{
Reverse codes variables
}
\usage{
revcode(data, vars)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
indicates your dataset.
}
  \item{vars}{
indicates the variable number or name to reverse code.
}
}

\author{
Francis Huang
}


\seealso{
\link[gendata:genmvnorm]{genmvnorm}
\link[gendata:dtrans]{dtrans}
\link[gendata:recalib]{recalib}
}

\keyword{ reverse }
\keyword{ recode }% __ONLY ONE__ keyword per line
